requirements.txt
discord.py == 1.2.5
urllib3 == 1.25.3